move framing to protocol
